Update release script and versioning; add RedisInsight functionality
