+++ VERSION
@@ -1,1 +1,1 @@
-2.8.11-stable
+2.8.12-beta
